move params.set to the MenuTeamSection
